+++ Cosine-Transform/main.h
@@ -1,6 +1,6 @@
 #include <stdlib.h>
 
 extern "C"{
-    void Cosine_transform(int* img, double* img_out);
-    void Inverse_cosine_transform(double* f_in, int* img);
+    void Cosine_transform(double* img, double* img_out);
+    void Inverse_cosine_transform(double* f_in, double* img);
 }
